Merged the deleteSkillRequest.js file with getrequest.js.
Finished resolving the issue with the deletebtn function not working when clicking the individual deletebtns. Had to assign the id as a parameter for the function and create and external onclick event rather than inline.

Also adding confirmation messages to the delete button to make sure the user wants to delete the skill associated with that id.